+++ internal/parse/parse_test.go
@@ -18,12 +18,24 @@ func init() {
 func TestParse(t *testing.T) {
 	ctx := t.Context()
 
-	info, err := PrimaryPackage(ctx, "go", "./testdata", []string{"func.go", "command.go", "alias.go", "repeating_synopsis.go", "subcommands.go"})
+	info, err := PrimaryPackage(ctx, "go", "./testdata", []string{"func.go", "command.go", "alias.go", "repeating_synopsis.go", "subcommands.go", "watch.go"})
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	expected := []Function{
+		{
+			Name:    "WatchTarget",
+			IsWatch: true,
+		},
+		{
+			Name:    "NonWatchTarget",
+			IsWatch: false,
+		},
+		{
+			Name:    "WatchDepsTarget",
+			IsWatch: true,
+		},
 		{
 			Name:     "ReturnsNilError",
 			IsError:  true,